requirements.txt
alembic==0.9.9
amqp==2.3.2
aniso8601==3.0.0
asn1crypto==0.24.0
attrs==18.1.0
Automat==0.6.0
backports-abc==0.5
bcrypt==3.1.4
beautifulsoup4==4.6.0
billiard==3.5.0.3
blinker==1.4
celery==4.1.1
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
constantly==15.1.0
crypto==1.4.1
cryptography==2.2.2
cssmin==0.2.0
cssselect==1.0.3
Django==1.11.12
django-haystack==2.8.1
django-redis-cache==1.7.1
django-tinymce==2.7.0
ecdsa==0.13
enum34==1.1.6
fabric==2.0.1
Flask==1.0.2
Flask-Admin==1.5.1
Flask-Assets==0.12
Flask-Bcrypt==0.7.1
Flask-Cache==0.13.1
Flask-Celery-Helper==1.1.0
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-OAuth==0.12
Flask-Principal==0.4.0
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
futures==3.2.0
httpie==0.9.9
httplib2==0.11.3
hyperlink==18.0.0
idna==2.6
image==1.5.24
incremental==17.5.0
invoke==1.0.0
ipaddress==1.0.22
itsdangerous==0.24
jieba==0.39
Jinja2==2.10
jsmin==2.2.2
kombu==4.2.1
lxml==4.2.1
Mako==1.0.7
MarkupSafe==1.0
meld3==1.0.2
mpmath==1.0.0
MySQL-python==1.2.5
Naked==0.1.31
oauth2==1.9.0.post1
paramiko==2.4.1
parsel==1.4.0
Pillow==5.1.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
pycrypto==2.6.1
PyDispatcher==2.0.5
pygame==1.9.3
Pygments==2.2.0
pymongo==3.6.1
PyMySQL==0.8.0
PyNaCl==1.2.1
pyOpenSSL==17.5.0
Pypubsub==4.0.0
python-dateutil==2.7.3
python-editor==1.0.3
pytz==2018.4
PyYAML==3.12
queuelib==1.5.0
redis==2.10.6
requests==2.18.4
rethinkdb==2.3.0.post6
Scrapy==1.5.0
scrapy-redis==0.6.8
selenium==3.11.0
service-identity==17.0.0
shellescape==3.4.1
singledispatch==3.4.0.3
six==1.11.0
SQLAlchemy==1.2.8
supervisor==3.3.4
sympy==1.1.1
tornado==5.0.2
torndb==0.3
Twisted==18.4.0
typing==3.6.4
urllib3==1.22
uWSGI==2.0.17
vine==1.1.4
w3lib==1.19.0
webassets==0.12.1
Werkzeug==0.14.1
Whoosh==2.7.4
WTForms==2.1
xmltodict==0.11.0
zope.interface==4.5.0
